improve display and persistence
DAO explicitly orders results by timestamp, rather than an id field.  So
the id can be freed from needing semantic meaning it its string format,
and is now a UUID.  DAO no longer directly emulates the Visit properties
of having an arrival and departure time since only one was used, and so
it is misleading to have both properties.  Instead it has a single
timestamp and boolean indicating if it is an arrival (or departure).

Changed the date/time format string to be slightly shorter for display
brevity.