add Bluetooth permission request
Significant rework of permissions request flow.  The Main fragment now
performs an initial request for all permissions on the first start, when
Android determines that a rationale does not need to be presented to the
user.  If any permissions are not granted, the app will proceed to
request them individually after providing a rationale to the user in the
Permission fragment.

Insert Bluetooth permission request in the above flow.

Refactor starting and configuring the AirshipAdapter to a new Gimbal
Integration class so it may be invoked from both Main and Permission
fragments.