Add sample app (#5)
* create sample app

* refactor sample app startup behavior

Be ready to receive Airship adapter place events at app start time.

Call adapter's `start()` just once, upon permissions grant.  Allow the
adapter to restore operation instead of explicitly calling restore and
start.  Enable user push notifications, also just once.

* improve display and persistence

DAO explicitly orders results by timestamp, rather than an id field.  So
the id can be freed from needing semantic meaning it its string format,
and is now a UUID.  DAO no longer directly emulates the Visit properties
of having an arrival and departure time since only one was used, and so
it is misleading to have both properties.  Instead it has a single
timestamp and boolean indicating if it is an arrival (or departure).

Changed the date/time format string to be slightly shorter for display
brevity.

* request both coarse and fine permissions

Android says that both permissions must be requested if fine is
desired.  It seemed to work okay just requesting fine, but...

* fix spurious error in Android Studio editor

Use Kotlin version of Gradle plugin to preprocessors do their magic so
that AS doesn't display errors that don't really exist.

* fix transcript event date format

It was displaying the 12 hour format but without AM/PM. Now it's 24-hour
format.

* fix casting warning

Somehow it built okay but Android Studio was showing an error in the
editor.  It should be casted anyways.

* initialize Airship at app start

Use default Airship-provided AutoPilot to use
`airshipconfig.properties` asset-based configuration.  The call to
`UAirship.takeOff()` was in `MainFragment` which meant that background
app-waking location events would be dropped, as Airship would not be
initialized.

* fix permissions granted behavior

Call onPermissionsGranted when permissions are granted, not as part of
the main fragment creation.  Progress through the permissions requests
successfully on older versions of Android (pre-notification-permission).

The Gimbal API key is now set in `PermissionViewModel.kt`.

* add Bluetooth permission request

Significant rework of permissions request flow.  The Main fragment now
performs an initial request for all permissions on the first start, when
Android determines that a rationale does not need to be presented to the
user.  If any permissions are not granted, the app will proceed to
request them individually after providing a rationale to the user in the
Permission fragment.

Insert Bluetooth permission request in the above flow.

Refactor starting and configuring the AirshipAdapter to a new Gimbal
Integration class so it may be invoked from both Main and Permission
fragments.

* add background location permission request

* add state indicators to Sample app UI

* add current permissions indicators
* add switch to enable/disable Gimbal and indicate current state.
  Gimbal is enabled automatically after FINE location is granted by the
  user, either before or after display of rationale to the user.

* add sample app setup to README

Also more information about `CustomEvent` properties.

---------

Co-authored-by: Michael <michael.vance@gimbal.com>